Added boot screen and seq
