+++ app/src/hw_module.c
@@ -144,6 +144,10 @@ static const struct battery_model battery_model = {
 /* nPM1300 CHARGER.BCHGCHARGESTATUS.CONSTANTCURRENT register bitmask */
 #define NPM1300_CHG_STATUS_CC_MASK BIT_MASK(3)
 
+extern struct k_sem sem_disp_ready;
+
+extern struct k_msgq q_disp_boot_msg;
+
 static void gpio_keys_cb_handler(struct input_event *evt)
 {
     // printk("GPIO_KEY %s pressed, zephyr_code=%u, value=%d type=%d\n",
@@ -589,6 +593,18 @@ static void event_callback(const struct device *dev, struct gpio_callback *cb, u
     }
 }
 
+static void hw_add_boot_msg(char *msg, bool status)
+{
+    struct hpi_boot_msg_t boot_msg = {
+        
+        .status = status,
+    };
+    strcpy(boot_msg.msg, msg);
+
+    k_msgq_put(&q_disp_boot_msg, &boot_msg, K_NO_WAIT);
+
+}
+
 void hw_init(void)
 {
     int ret = 0;
@@ -601,9 +617,12 @@ void hw_init(void)
 
     k_sem_give(&sem_display_on);
 
+    k_sem_take(&sem_disp_ready, K_FOREVER);
+
     if (!device_is_ready(pmic))
     {
         LOG_ERR("Pmic device not ready");
+        
         return 0;
     }
 
@@ -613,6 +632,8 @@ void hw_init(void)
         // return 0;
     }
 
+    //sh8601_reinit(display_dev);
+
     if (!device_is_ready(charger))
     {
         LOG_ERR("Charger device not ready.\n");
@@ -621,7 +642,10 @@ void hw_init(void)
     if (npm_fuel_gauge_init(charger) < 0)
     {
         LOG_ERR("Could not initialise fuel gauge.\n");
+        hw_add_boot_msg("PMIC", false);
         // return 0;
+    } else {
+        hw_add_boot_msg("PMIC", true);
     }
 
     static struct gpio_callback event_cb;
@@ -651,9 +675,12 @@ void hw_init(void)
     if (!device_is_ready(imu_dev))
     {
         LOG_ERR("Error: Accelerometer device not ready");
+        hw_add_boot_msg("BMI323", false);
+
     }
     else
     {
+        hw_add_boot_msg("BMI323", true);
         if (sensor_trigger_set(imu_dev, &imu_trig, trigger_handler) < 0)
         {
             LOG_ERR("Could not set trigger");
@@ -669,14 +696,14 @@ void hw_init(void)
         }
     }
 
-    sh8601_reinit(display_dev);
+    
 
     // device_init(display_dev);
     //  k_sleep(K_MSEC(1000));
     // hw_pwr_display_enable();
 
-    // regulator_enable(ldsw_sens_1_8);
-    regulator_disable(ldsw_sens_1_8);
+    regulator_enable(ldsw_sens_1_8);
+    //regulator_disable(ldsw_sens_1_8);
 
     k_sleep(K_MSEC(100));
 
@@ -696,9 +723,11 @@ void hw_init(void)
     {
         LOG_ERR("MAX30001 device not found!");
         max30001_device_present = false;
+        hw_add_boot_msg("MAX30001", false);
     }
     else
     {
+        hw_add_boot_msg("MAX30001", true);
         LOG_INF("MAX30001 device found!");
         max30001_device_present = true;
 
@@ -724,9 +753,11 @@ void hw_init(void)
     {
         LOG_ERR("MAX32664C device not present!");
         max32664c_device_present = false;
+        hw_add_boot_msg("MAX32664C", false);
     }
     else
     {
+        hw_add_boot_msg("MAX32664C", true);
         LOG_INF("MAX32664C device present!");
         max32664c_device_present = true;
 
@@ -759,11 +790,13 @@ void hw_init(void)
     {
         LOG_ERR("MAX32664D device not present!");
         max32664d_device_present = false;
+        hw_add_boot_msg("MAX32664D", false);
     }
     else
     {
         LOG_INF("MAX32664D device present!");
         max32664d_device_present = true;
+        hw_add_boot_msg("MAX32664D", true);
 
         struct sensor_value mode_set;
 
@@ -781,10 +814,12 @@ void hw_init(void)
     if (!device_is_ready(max30208_dev))
     {
         LOG_ERR("MAX30208 device not found!");
+        hw_add_boot_msg("MAX30208", false);
     }
     else
     {
         LOG_INF("MAX30208 device found!");
+        hw_add_boot_msg("MAX30208", true);
     }
 
 
@@ -844,6 +879,8 @@ static uint32_t acc_get_steps(void)
     sensor_channel_get(imu_dev, SENSOR_CHAN_ACCEL_X, &steps);
     return (uint32_t)steps.val1;
 }
+
+
 
 void hw_thread(void)
 {
